config module, climb to n1
I've introduced a module with the logger and yaml configuration to be able to call the logger for debuggin from the different modules in the project. I've also continued working on trying to describe the positions that a song climbs from the lowest position in the charts to the n1 position. I've calculated the bottom position prior to  the song becoming n1 and the date of the position. I now intend to do a cumulative distribution function to describe the distribution for the complete time period and each decade.