Refactoring for cdf functions
I've carried out a refactoring of the code. I now use a summary of the n1 songs to which I ping different properties. I've also changed the cdf functions to be able to graph the different properties by using them as an argument in the cdf functions.

I also filter the songs that might have had a re-entry in the charts so the time that it takes for the song to climb from bottom to n1 is accurate for the specific time period of that stint of that song in the charts.