Bottom pos calculation changes
I've changed the calculations of the bottom position in charts to get positions that have been parsed prior to the song reaching its peak position. This way I'll be able to calculate the true climb of the song from the lowest position prior to reaching its peak to its peak.